Better compression for codec: 2x.
git-svn-id: file:///home/gcadet/Downloads/v6_svn/v6/trunk@204 435ad755-7caf-5a4d-9ef7-76b6bf021c60

+++ source/v6/core/memory.h
@@ -13,7 +13,7 @@ public:
 	virtual void *	alloc( u64 nSize ) = 0;
 
 	template <typename T>
-	T *				newArray( u32 count )
+	T *				newArray( u64 count )
 	{
 		return (T*)alloc( sizeof(T) * count );
 	}
